+++ src/routes/JobCreateRoute.js
@@ -1,20 +1,28 @@
 import React from 'react';
 import compose from 'compose-function';
 import PropTypes from 'prop-types';
-
 import { stripesConnect } from '@folio/stripes/core';
+import withFileHandlers from './components/withFileHandlers';
+import View from '../components/views/JobForm';
 
 class JobCreateRoute extends React.Component {
   static manifest = Object.freeze({
     jobs: {
       type: 'okapi',
-      path: 'erm/jobs',
+      path: 'erm/jobs/packageImport',
       fetch: false,
       shouldRefresh: () => false,
     },
   });
 
   static propTypes = {
+    handlers: PropTypes.object,
+    history: PropTypes.shape({
+      push: PropTypes.func.isRequired,
+    }).isRequired,
+    location: PropTypes.shape({
+      search: PropTypes.string.isRequired,
+    }).isRequired,
     mutator: PropTypes.shape({
       jobs: PropTypes.shape({
         POST: PropTypes.func.isRequired,
@@ -22,15 +30,41 @@ class JobCreateRoute extends React.Component {
     }).isRequired,
   };
 
+  static defaultProps = {
+    handlers: {},
+  }
+
+  handleClose = () => {
+    const { location } = this.props;
+    this.props.history.push(`/local-kb-admin${location.search}`);
+  }
+
+  handleSubmit = (job) => {
+    const { history, location, mutator } = this.props;
+
+    mutator.jobs
+      .POST(job)
+      .then(({ id }) => {
+        history.push(`/local-kb-admin/${id}${location.search}`);
+      });
+  }
+
   render() {
+    const { handlers } = this.props;
+
     return (
-      <div>
-        Create page
-      </div>
+      <View
+        handlers={{
+          ...handlers,
+          onClose: this.handleClose
+        }}
+        onSubmit={this.handleSubmit}
+      />
     );
   }
 }
 
 export default compose(
+  withFileHandlers,
   stripesConnect
 )(JobCreateRoute);