+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change history for ui-local-kb-admin
 
 ## 1.1.0 IN PROGRESS
+* Added functionality to create jobs. ERM-337
 * Added settings page for External Data Sources. ERM-83
 
 ## 1.0.0 2019-07-22